Changed place for checking if the order is paid. Everything works.
